Update on "dump operator names of a module and its sub-modules."
