+++ torch/csrc/jit/export.cpp
@@ -917,32 +917,6 @@ std::string prettyPrint(const onnx::ModelProto& model) {
 
 } // namespace
 
-void dump_opnames(const script::Module& m, std::unordered_set<std::string>& opnames) {
-  auto methods = m.get_methods();
-  for (const auto &method : methods) {
-    const auto &func = method.function();
-//    std::cout << "function name: " << func.name() << std::endl;
-    torch::jit::Code code(func.graph());
-    for (size_t i = 0; i < code.instructions().size(); ++i) {
-      auto ins = code.instructions()[i];
-      auto node = code.instructions_source()[i];
-      if (ins.op == OpCode::OP) {
-        auto opname = node->schema().operator_name();
-        std::string namestr = opname.name;
-        if (!opname.overload_name.empty())
-          namestr += opname.overload_name;
-//        std::cout << "    " << namestr << std::endl;
-        opnames.emplace(namestr);
-      }
-    }
-  }
-  auto modules = m.get_modules();
-  for (const auto& sub_m : modules) {
-//    std::cout << "sub module name: " << sub_m.name << std::endl;
-    dump_opnames(sub_m.module, opnames);
-  }
-}
-
 void SetExportModuleExtraFilesHook(ExportModuleExtraFilesHook hook) {
   GetExtraFilesHook() = hook;
 }
